backend/django/requirements.txt
django ~= 4.0
authlib ~= 1.0
python-dotenv ~= 0.19
requests ~= 2.27
psycopg2
#pymongo ~= 3.6
#dnspython
#mongoengine
#djongo
#pytz
#django-utils-six
